Issue #ED-2226 merge: Rel 7.0.0 min env (#8959)
* ED-2226 fix: updated the uci-console package

* removed non-mand env variable

* #ED-2732 : minimized with 40 env variable

* ED-2720 feat: minimized env variable to 38

* ED-2720 feat: minimized env variable to 31

* ED-2720 feat: minimized env variable to 31

* ED-2720 feat: minimized env variable to 31

* ED-2720 feat: fix the feedback comment

* Segragate Mandatory and Optional Environmental Variables

* Segragate Mandatory Values and Optional Values

* Mandatory Default Values changes in OptionalEnv

* mandatory default Segregation

* Mandatory Default and Optional Values

* Added Safety Checks

* Readme Update: Mandatory Environmental variables added

* File path is segregated as mandatory and optional

* Readme Update: Optional env variables are removed and confluence link for further reference are added

* reference added as a clickable link

* Readme Update: description of reference link updated

---------
