+++ README.md
@@ -108,46 +108,18 @@ Installing Sunbird requires two primary software components:
     
 
 4. Edit the Application Configuration
+    > Open `<PROJECT-FOLDER>/src/app/helpers/mandatoryEnv.js` in any available text editor with the below keys along with the values as per the infra
+                                                                                                                                                  
+          |       Variable Name                     |                   Description                                       |
+          | :-------------------------------------- |---------------------------------------------------------------------|
+          |  sunbird_anonymous_register_token       |   Token to register anonymous device                                |                                  
+          |  sunbird_anonymous_default_token        |   Fallback token for device register API for `anonymous` users      |                                                           
+          |  sunbird_loggedin_device_register_api   |   Kong End Point for Logged Users                                   |                               
+          |  sunbird_logged_default_token           |   Fallback token for device register API for `logged-in` users      |                                                         
+          |  sunbird_loggedin_register_token        |   Device register API for logged-in users                           | 
+          |  cloud_private_storage_secret           |   To set the Cloud Account Key                                      |
 
-    > These are the mandatory keys required to run the application in Local environment. Please update them with appropriatte values in `<PROJECT-FOLDER>/src/app/helpers/environmentVariablesHelper.js`
-
-        |           Environment Variable        |  Data Type |             Description                |
-        | :-------------------------------------| ---------- | -------------------------------------  |
-        |     sunbird_cloud_storage_provider    |   string   |        Cloud Service Provider          |
-        |   cloud_private_storage_accountname   |   string   |          Cloud Account Name            |
-        |      cloud_private_storage_secret     |   string   |          Cloud Account Key             |
-        |  KONG_DEVICE_REGISTER_ANONYMOUS_TOKEN |   boolean  |   Flag value to allow anonymous user   |
-        |  sunbird_anonymous_device_register_api|   string   |The API for registering anonymous device|
-        |  sunbird_anonymous_register_token     |   string   |    Token to register anonymous device  |
-        |               SB_DOMAIN               |   string   |    The host for Sunbird Environment    |
-        |         PORTAL_API_AUTH_TOKEN         |   string   |     User generated API auth token      |
-
-
-    > Open `<PROJECT-FOLDER>/src/app/helpers/environmentVariablesHelper.js` in any available text editor and update the contents of the file so that it contains exactly the following values
-
-      ```console
-          module.exports = {
-              // 1. LEARNER_URL   
-              LEARNER_URL: env.sunbird_learner_player_url || <'https://<host for adopter's instance>',
-              
-              // 2. CONTENT_URL
-              CONTENT_URL: env.sunbird_content_player_url || <'https://<host for adopter's instance>',
-              
-              // 3. CONTENT_PROXY  
-              CONTENT_PROXY_URL: env.sunbird_content_proxy_url || <'https://<host for adopter's instance>',
-              PORTAL_REALM: env.sunbird_portal_realm || 'sunbird',
-              
-              // 4. PORTAL_AUTH_SERVER_URL
-              PORTAL_AUTH_SERVER_URL: env.sunbird_portal_auth_server_url || <'https://<host for adopter's instance>',
-              PORTAL_AUTH_SERVER_CLIENT: env.sunbird_portal_auth_server_client || "portal",
-              ...
-              PORTAL_PORT: env.sunbird_port || 3000,
-                
-              // 5. PORTAL_ECHO_API_URL
-              PORTAL_ECHO_API_URL: env.sunbird_echo_api_url || '',
-              ...
-          }
-      ```
+    > For further environment variable reference refer to this confluence wiki link: [https://project-sunbird.atlassian.net/wiki/spaces/SP/pages/3353378817/Portal+-+Min+environment+variables](https://project-sunbird.atlassian.net/wiki/spaces/SP/pages/3353378817/Portal+-+Min+environment+variables)
 
     > Once the file is updated with appropriate values, then you can proceed with running the application
 
@@ -219,3 +191,4 @@ Installing Sunbird requires two primary software components:
 | [@project-sunbird/sb-notification](https://www.npmjs.com/package/@project-sunbird/sb-notification)           	| https://github.com/Sunbird-Ed/sb-notification                    	| 6.0.0       	| NG 14      	|
 | [@shikshalokam/sl-questionnaire](https://www.npmjs.com/package/@shikshalokam/sl-questionnaire)           	    | https://github.com/shikshalokam/sl-questionnaire-components       | 2.3.0       	| NG 12      	|
 | [@shikshalokam/sl-reports-library](https://www.npmjs.com/package/@shikshalokam/sl-reports-library)           	| https://github.com/shikshalokam/sl-reports-library                | 3.0.1       	| NG 14      	|
+
