+++ src/app/framework.config.js
@@ -37,7 +37,7 @@ module.exports = {
         uid: uuidv1(),
         endpoint: telemtryEventConfig.endpoint,
         batchsize: 1,
-        host: envHelper.TELEMETRY_SERVICE_LOCAL_URL,
+        host: envHelper?.TELEMETRY_SERVICE_LOCAL_URL,
         authtoken: 'Bearer ' + envHelper.PORTAL_API_AUTH_TOKEN,
         runningEnv: 'server'
     },
