+++ src/app/helpers/cloudStorage/index.js
@@ -11,7 +11,7 @@
 
 const cloudService  = require('client-cloud-services');
 const envHelper     = require('../../helpers/environmentVariablesHelper');
-const cloudProvider = envHelper.sunbird_cloud_storage_provider;
+const cloudProvider = envHelper?.sunbird_cloud_storage_provider;
 
 /**
  * Based on Environment Cloud Provider value
@@ -19,14 +19,14 @@ const cloudProvider = envHelper.sunbird_cloud_storage_provider;
  */
 if (!cloudProvider) throw new Error("Cloud Storage Service - Provider is not initialized");
 let cloudConfig = {
-  provider: envHelper.sunbird_cloud_storage_provider,
-  identity: envHelper.cloud_private_storage_accountname,
-  credential: envHelper.cloud_private_storage_secret,
-  privateObjectStorage: envHelper.cloud_storage_privatereports_bucketname,
-  publicObjectStorage: envHelper.cloud_storage_resourceBundle_bucketname,
-  region: envHelper.cloud_private_storage_region,
-  projectId: envHelper.cloud_private_storage_project,
-  endpoint:envHelper.cloud_private_storage_endpoint
+  provider: envHelper?.sunbird_cloud_storage_provider,
+  identity: envHelper?.cloud_private_storage_accountname,
+  credential: envHelper?.cloud_private_storage_secret,
+  privateObjectStorage: envHelper?.cloud_storage_privatereports_bucketname,
+  publicObjectStorage: envHelper?.cloud_storage_resourceBundle_bucketname,
+  region: envHelper?.cloud_private_storage_region,
+  projectId: envHelper?.cloud_private_storage_project,
+  endpoint:envHelper?.cloud_private_storage_endpoint
 };
 
 let cloudClient = cloudService.init(cloudConfig);
