+++ src/app/helpers/constants.js
@@ -1,4 +1,5 @@
 /*jslint node: true, nomen: true, plusplus: true*/
+const SB_DOMAIN = require('./environmentVariablesHelper').SB_DOMAIN;
 'use strict';
 
 module.exports = {
@@ -18,5 +19,6 @@ module.exports = {
     V1: 'v1',
     V3: 'v3'
   },
+  LOCALHOST:'http://localhost:3000/',
   GOOGLE_VERIFICATION_URL: 'https://www.google.com/recaptcha/api/siteverify'
 };
