+++ src/app/helpers/googleOauthHelper.js
@@ -14,81 +14,81 @@ const { ProxyLogger } = require("@project-sunbird/logger/decorator");
 const { getBearerToken } = require('../helpers/kongTokenHelper')
 
 const keycloakGoogle = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.secret
   }
 })
 
 // keycloack client for account merge poiting to subdomain
 const keycloakMergeGoogle = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_MERGE_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_MERGE_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.secret
   }
 })
 
 
 const keycloakGoogleAndroid = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_ANDROID_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_ANDROID_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_ANDROID_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_ANDROID_CLIENT?.secret
   }
 })
 const keycloakMergeGoogleAndroid = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_ANDROID_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_ANDROID_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_MERGE_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_MERGE_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_ANDROID_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_ANDROID_CLIENT?.secret
   }
 })
 const keycloakGoogleIos = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_IOS_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_IOS_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_MERGE_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_MERGE_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_IOS_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_IOS_CLIENT?.secret
   }
 })
 
 const keycloakGoogleDesktop = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_DESKTOP_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_DESKTOP_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_AUTH_SERVER_URL,
+  serverUrl: envHelper?.PORTAL_AUTH_SERVER_URL,
   realm: envHelper.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_DESKTOP_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_DESKTOP_CLIENT?.secret
   }
 })
 const keycloakMergeGoogleDesktop = getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_DESKTOP_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_DESKTOP_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_MERGE_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_MERGE_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_DESKTOP_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_DESKTOP_CLIENT?.secret
   }
 })
 
 const keycloakMergeGoogleIos =  getKeyCloakClient({
-  resource: envHelper.KEYCLOAK_GOOGLE_CLIENT.clientId,
+  resource: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.clientId,
   bearerOnly: true,
-  serverUrl: envHelper.PORTAL_MERGE_AUTH_SERVER_URL,
-  realm: envHelper.PORTAL_REALM,
+  serverUrl: envHelper?.PORTAL_MERGE_AUTH_SERVER_URL,
+  realm: envHelper?.PORTAL_REALM,
   credentials: {
-    secret: envHelper.KEYCLOAK_GOOGLE_CLIENT.secret
+    secret: envHelper?.KEYCLOAK_GOOGLE_CLIENT?.secret
   }
 })
