+++ src/app/helpers/utils.js
@@ -1,0 +1,7 @@
+const CONSTANTS = require("../helpers/constants");
+const envVariables = require('./environmentVariablesHelper');
+
+function defaultHost(envValue) {
+    return envValue || CONSTANTS.LOCALHOST;
+}
+module.exports = { defaultHost, envVariables }
