+++ src/app/proxy/localProxy.js
@@ -1,7 +1,7 @@
 const proxy = require('express-http-proxy')
-const envHelper = require('./../helpers/environmentVariablesHelper.js')
-const learnerBaseUrl = envHelper.LEARNER_SERVICE_UPSTREAM_URL
-const contentBaseUrl = envHelper.CONTENT_SERVICE_UPSTREAM_URL
+const utils = require('../helpers/utils.js');
+const learnerBaseUrl  = utils?.defaultHost(utils?.envVariables?.LEARNER_SERVICE_UPSTREAM_URL);
+const contentBaseUrl  = utils?.defaultHost(utils?.envVariables?.CONTENT_SERVICE_UPSTREAM_URL);
 const csLocalProxyURI = '/localproxy/content/'
 const lsLocalProxyURI = '/localproxy/learner/'
 const reqDataLimitOfContentUpload = '30mb'
