+++ src/app/routes/accountRecoveryRoute.js
@@ -5,17 +5,19 @@ const dateFormat = require('dateformat');
 const uuidv1 = require('uuid/v1');
 const proxy = require('express-http-proxy');
 const proxyUtils = require('../proxy/proxyUtils.js');
+const utils = require('../helpers/utils.js');
 const { logger } = require('@project-sunbird/logger');
 const { encriptWithTime, encrypt } = require('../helpers/crypto');
 const { decodeNChkTime } = require('../helpers/utilityService');
 const googleService = require('../helpers/googleService');
-
+const learner_Service_Local_BaseUrl = utils?.defaultHost(utils?.envVariables?.learner_Service_Local_BaseUrl);
+const LEARNER_URL  = utils?.defaultHost(utils?.envVariables?.LEARNER_URL);
 module.exports = (app) => {
 
   app.post('/learner/user/v1/fuzzy/search',
   googleService.validateRecaptcha,
-  proxy(envHelper.learner_Service_Local_BaseUrl, {
-    proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(envHelper.learner_Service_Local_BaseUrl),
+  proxy(learner_Service_Local_BaseUrl, {
+    proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(learner_Service_Local_BaseUrl),
     proxyReqPathResolver: (req) => {
       logger.info({ msg: `${req.url} called`});
     return '/private/user/v1/search';
@@ -50,8 +52,8 @@ module.exports = (app) => {
         });
       }
     },
-    proxy(envHelper.learner_Service_Local_BaseUrl, {
-      proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(envHelper.learner_Service_Local_BaseUrl),
+    proxy(learner_Service_Local_BaseUrl, {
+      proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(learner_Service_Local_BaseUrl),
       proxyReqPathResolver: (req) => {
         return '/private/user/v1/password/reset'; // /private/user/v1/reset/password
       }
@@ -59,10 +61,10 @@ module.exports = (app) => {
 
   app.all('/learner/otp/v1/verify',
     bodyParser.urlencoded({ extended: false }), bodyParser.json({ limit: '10mb' }),
-    proxy(envHelper.LEARNER_URL, {
-      proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(envHelper.LEARNER_URL),
+    proxy(LEARNER_URL, {
+      proxyReqOptDecorator: proxyUtils.decorateRequestHeaders(LEARNER_URL),
       proxyReqPathResolver: (req) => {
-        return require('url').parse(envHelper.LEARNER_URL + req.originalUrl.replace('/learner/', '')).path
+        return require('url').parse(LEARNER_URL + req.originalUrl.replace('/learner/', '')).path
       },
       userResDecorator: (proxyRes, proxyResData, req, res) => {
         try {
