+++ src/app/routes/groupRoutes.js
@@ -1,6 +1,7 @@
 const proxyUtils = require('../proxy/proxyUtils.js')
 const envHelper = require('../helpers/environmentVariablesHelper.js')
-const learnerURL = envHelper.LEARNER_URL
+const utils = require('../helpers/utils.js');
+const learnerURL  = utils?.defaultHost(utils?.envVariables?.LEARNER_URL);
 const proxy = require('express-http-proxy')
 const reqDataLimitOfContentUpload = '50mb'
 const Telemetry = require('../libs/sb_telemetry_util/telemetryService.js')
