+++ src/app/routes/ssoRoutes.js
@@ -12,7 +12,7 @@ const {generateAuthToken, getGrantFromCode} = require('../helpers/keyCloakHelper
 const {parseJson, isDateExpired} = require('../helpers/utilityService');
 const {getUserIdFromToken} = require('../helpers/jwtHelper');
 const fs = require('fs');
-const externalKey = envHelper.CRYPTO_ENCRYPTION_KEY_EXTERNAL;
+const externalKey = envHelper?.CRYPTO_ENCRYPTION_KEY_EXTERNAL;
 const successUrl = '/sso/sign-in/success';
 const updateContactUrl = '/sign-in/sso/update/contact';
 const errorUrl = '/sso/sign-in/error';
